+++ sematic/versions.py
@@ -5,7 +5,7 @@ import typing
 # the sdk. Should be bumped any time a release is made. Should be set
 # to whatever is the version after the most recent one in changelog.md,
 # as well as the version for the sematic wheel in wheel_version.bzl
-CURRENT_VERSION = (0, 12, 0)
+CURRENT_VERSION = (0, 13, 0)
 
 # Represents the smallest client version that works with the server
 # at the CURRENT_VERSION. Should be updated any time a breaking change