+++ docs/changelog.md
@@ -4,6 +4,10 @@ Lines for version numbers should always be formatted as `* MAJOR.MINOR.PATCH`
 with nothing else on the line.
 -->
 * HEAD
+* 0.13.0
+    * [bugfix] Bugfix for dicitonary visualization
+    * [feature] When running in cloud mode, have the server log to stdout
+    * [feature] Enable mounting Kubernetes secrets into the container
 * 0.12.0
     * [feature] BREAKING CHANGE: Allow specifying resource requirements for Kubernetes
     jobs. `KubernetesResourceRequirements` has a new required field, `requests`